Make raw page data generator for DPLUtils unit tests a separate class
There will be more unit tests and micro benchmarks using the data
generator.
